requirements.txt
Django
djangorestframework
django-storages
boto3
python-dotenv
pytz
simplejson
unicodecsv
argparse